ADDED: python module to actually listen joint state and trigger pyqt signal-slot for gui
yea, so, package

but it's ok, coz, to actually make the GUI able to listen to the joint_states topic, a python module on the same folder as the gui node needs to be created. it just made the file more modular (technically, we can put the code inside the GUI node).